requirements.txt
fastapi[standard]==0.128.0
pydantic[email]==2.12.5
sqlalchemy==2.0.46
alembic==1.18.2
psycopg2-binary==2.9.10
asyncpg==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.20
Jinja2==3.1.5
email-validator==2.2.0
httpx==0.28.1
packaging==25.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2